pyproject.toml
[build-system]
requires = ["setuptools>=69", "setuptools_scm[toml]>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "splitweave"
dynamic = ["version"]
description = "A DSL for 2D pattern programs, for synthetic pattern and analogy generation."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
  { name = "Aditya Ganeshan", email = "adityaganeshan@gmail.com" }
]
requires-python = ">=3.10"
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering",
]
dependencies = [
  "numpy>=1.23",
  "sympy>=1.12",
  "torch>=2.0",
  "yacs>=0.1.8",
  "Pillow>=9.0",
  "geolipi>=0.1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "build>=1.2.0",
  "ruff>=0.5.0",
  "black>=24.0",
  "mypy>=1.0",
]

[project.urls]
Homepage = "https://github.com/BardOfCodes/splitweave"
Repository = "https://github.com/BardOfCodes/splitweave"
Issues = "https://github.com/BardOfCodes/splitweave/issues"

[tool.setuptools.packages.find]
include = ["splitweave*"]

[tool.setuptools_scm]
version_scheme = "guess-next-dev"
local_scheme = "node-and-date"
fallback_version = "0.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py310"
src = ["splitweave", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.ruff_cache
    | \.venv
    | _build
    | build
    | dist
)/
'''

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
